media-source/mediasource-htmlmediaelement-capturestream.html

<!DOCTYPE html>
<html>
<head>
  <title>Test HTMLMediaElement.captureStream() with src set to MediaSource Blob URL</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body>
  <p id="test">click</p>
  <script>
    function testMediaSourceMediaRecorder() {
      return new Promise(function(resolve, reject) {
        const test = document.getElementById("test");
        const captureStream = mediaElement =>
          !!mediaElement.mozCaptureStream ? mediaElement.mozCaptureStream() : mediaElement.captureStream();
        class MediaFragmentRecorder {
          constructor({
            urls = [],
              video = document.createElement("video"),
              width = 320,
              height = 280
          } = {}) {
            if (urls.length === 0) {
              throw new TypeError("no urls passed to MediaFragmentRecorder");
            }
            return (async() => {
              video.height = height;
              video.width = width;
              video.autoplay = true;
              video.preload = "auto";
              video.controls = true;
              video.muted = false;
              const chunks = [];
              let duration = 0;
              let media = await Promise.all(
                urls.map(async({
                  from,
                  to,
                  src
                }, index) => {
                  const url = new URL(src);
                  if (url.hash.length) {
                    [from, to] = url.hash.match(/\d+/g);
                  }
                  return {
                    blob: await fetch(src).then(response => response.blob()),
                    from,
                    to
                  }
                }));
              for (let {
                  from,
                  to,
                  blob
                }
                of media) {
                await new Promise(async(resolve) => {
                  let recorder;
                  const blobURL = URL.createObjectURL(blob);
                  video.addEventListener("playing", e => {
                    const mediaStream = captureStream(video);
                    recorder = new MediaRecorder(mediaStream, {
                      mimeType: "video/webm;codecs=vp8,opus"
                    });
                    recorder.start();
                    recorder.addEventListener("stop", e => {
                      resolve();
                    }, {
                      once: true
                    });
                    recorder.addEventListener("dataavailable", async(e) => {
                      chunks.push(await new Response(e.data).arrayBuffer());
                    });
                    video.addEventListener("pause", e => {
                      if (recorder.state === "recording") {
                        recorder.stop();
                      } else {
                        recorder.requestData();
                      }
                      duration += video.currentTime - from;
                    }, {
                      once: true
                    });
                  }, {
                    once: true
                  });
                  video.addEventListener("canplay", e => video.play(), {
                    once: true
                  });
                  video.src = `${blobURL}#t=${from},${to}`;
                })
              };
              video.load();
              return {
                chunks,
                width,
                height
              }
            })()
          }
        }
        let urls = [{
          src: "https://upload.wikimedia.org/wikipedia/commons/a/a4/Xacti-AC8EX-Sample_video-001.ogv",
          from: 0,
          to: 4
        }, {
          src: "https://mirrors.creativecommons.org/movingimages/webm/ScienceCommonsJesseDylan_240p.webm#t=10,20"
        }, {
          from: 55,
          to: 60,
          src: "https://nickdesaulniers.github.io/netfix/demo/frag_bunny.mp4"
        }];

        test.addEventListener("click", _ =>
          new MediaFragmentRecorder({
            urls
          })
          .then(async({
            chunks,
            width,
            height
          }) => {
              let mediaStream, sourceBuffer;
              const video = document.createElement("video");
              video.height = height;
              video.width = width;
              video.autoplay = true;
              video.preload = "auto";
              video.controls = true;
              document.body.appendChild(video);
              const mediaSource = new MediaSource();
              const mimeCodec = "video/webm;codecs=vp8,opus";
              const sourceOpen = e => {
                sourceBuffer = mediaSource.addSourceBuffer(mimeCodec);
                sourceBuffer.mode = "sequence";
                sourceBuffer.addEventListener("updateend", e => {
                  if (chunks.length) {
                    sourceBuffer.appendBuffer(chunks.shift())
                  } else {
                    sourceBuffer.abort();
                    mediaSource.endOfStream();
                  } 
                });
                if (chunks.length) {
                  sourceBuffer.appendBuffer(chunks.shift());
                }
              }
              const handleWaiting = e => {
                if (chunks.length) {
                  sourceBuffer.appendBuffer(chunks.shift());
                } else {
                  mediaSource.endOfStream();
                  video.removeEventListener("waiting", handleWaiting);
                }
              }
              video.addEventListener("playing", e => {
                mediaStream = captureStream(video);
                // should `addtrack` event be dispatched twice?
                mediaStream.addEventListener("addtrack", e => { console.log(e); if (mediaStream.getTracks().length > 1) { resolve(mediaStream) } });
                mediaStream.addEventListener("active", e => { console.log(e); });
              }, {once: true});
              video.addEventListener("waiting", handleWaiting);
              mediaSource.addEventListener("sourceopen", sourceOpen);
              video.src = URL.createObjectURL(mediaSource);
          })
        , {once: true})
      })
    }
    promise_test(function(t) {
      return testMediaSourceMediaRecorder().then(function(result) {
        assert_equals(result, new MediaStream(), "Browser tab should not momentarily crash")
      });
    }, "Test HTMLMediaElement.captureStream() with src set to MediaSource Blob URL");
  </script>
</body>
</html>
